requirements.txt
accelerate>=0.34.2
bitsandbytes>=0.43.3
datasets>=2.21.0
fastapi>=0.115.0
peft>=0.12.0
pydantic>=2.9.0
pyyaml>=6.0.2
sentencepiece>=0.2.0
torch>=2.3.1
transformers>=4.44.2
trl>=0.10.1
uvicorn[standard]>=0.30.6
vllm>=0.6.1
wandb>=0.17.8
